Tightened up sync logic to not dump reference file info when only syncing mappings.
git-svn-id: https://aeon.stsci.edu/ssb/svn/crds/trunk@1350 0d8f46c8-9b30-49d0-b470-0f6283dc92e8

+++ crds/sync.py
@@ -147,21 +147,21 @@ class SyncScript(cmdline.ContextsScript):
         self.require_server_connection()
         if self.contexts:
             active_mappings = self.get_context_mappings()
-            if self.args.datasets:
-                active_references = self.sync_datasets()
-            else:
-                active_references = self.get_context_references()
-            active_references += self.get_conjugates(active_references)
-            active_references = sorted(set(active_references))
-            if self.args.fetch_references:
-                self.fetch_references(active_references)
-            if self.args.purge_references:
-                self.purge_references(active_references)    
+            verify_file_list = active_mappings
+            if self.args.fetch_references or self.args.purge_references:
+                if self.args.datasets:
+                    active_references = self.sync_datasets()
+                else:
+                    active_references = self.get_context_references()
+                active_references += self.get_conjugates(active_references)
+                active_references = sorted(set(active_references))
+                if self.args.fetch_references:
+                    self.fetch_references(active_references)
+                    verify_file_list += active_references
+                if self.args.purge_references:
+                    self.purge_references(active_references)    
             if self.args.purge_mappings:
                 self.purge_mappings()
-            verify_file_list = active_mappings
-            if self.args.fetch_references:
-                verify_file_list += active_references
         elif self.args.files:
             self.sync_explicit_files()
             verify_file_list = self.args.files
